Backend/requirements.txt
attrs==19.3.0
autopep8==1.5.2
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
click==7.1.2
cryptography==2.9.2
fancycompleter==0.9.1
Flask==1.1.2
Flask-Admin==1.5.6
Flask-Cors==3.0.8
Flask-Login==0.5.0
Flask-SQLAlchemy==2.4.1
gunicorn==20.0.4
idna==2.9
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
more-itertools==8.2.0
msgpack-python==0.5.6
mypy==0.770
mypy-extensions==0.4.3
mysqlclient==1.4.6
packaging==20.3
Pillow==7.1.2
pluggy==0.13.1
py==1.8.1
pycodestyle==2.5.0
pycparser==2.20
pycurl==7.43.0
Pygments==2.6.1
PyMySQL==0.9.3
pyparsing==2.4.7
pyrepl==0.9.0
pytest==5.4.1
pytest-mypy==0.4.2
PyVimeo==1.0.11
remote-pdb==2.0.0
requests==2.23.0
six==1.10.0
SQLAlchemy==1.3.16
tinydb==4.1.0
tuspy==0.2.1
typed-ast==1.4.1
typing-extensions==3.7.4.2
urllib3==1.25.9
wcwidth==0.1.9
Werkzeug==1.0.1
wmctrl==0.3
WTForms==2.3.1
wheel==0.34.2
-e git://github.com/maxcountryman/flask-uploads.git#egg=flask_uploads
